Refactor `SystemContextSettingsTest`
