+++ server/src/main/java/io/spine/system/server/SystemSettings.java
@@ -164,6 +164,7 @@ public final class SystemSettings implements SystemFeatures {
      * @return self for method chaining
      * @see #forgetEvents()
      */
+    @CanIgnoreReturnValue
     public SystemSettings persistEvents() {
         this.storeEvents = true;
         return this;
@@ -177,6 +178,7 @@ public final class SystemSettings implements SystemFeatures {
      * @return self for method chaining
      * @see #persistEvents()
      */
+    @CanIgnoreReturnValue
     public SystemSettings forgetEvents() {
         this.storeEvents = false;
         return this;
